Added extra functions to Matrix class. Added in check for Qt 4.6 in Gui::initialise() to allow setting of QGL rendering type (to fix broken rendering otherwise. Removed Model::drawPixelWidth() function. FIxed - Back-projection of screen coordinates to model coordinates now working. Fixed - Default value of TCanvas::currentDrawDepth_ now takes on sensible value. Moved Aten::sketchElement_ to TCanvas. Added text label rendering to bond drawing. Fixed - RenderEngine::renderTextPrimitive(int,int,const char*) was rendering at inverse y-coordinate.

+++ branches/engine/src/render/engine.h
@@ -91,7 +91,7 @@ class RenderEngine
 	// Project given model coordinates into world coordinates (and screen coordinates if Vec3 is supplied)
 	Vec3<double> &modelToWorld(Vec3<double> &pos, Vec4<double> *screenr = NULL, double screenradius = 0.0);
 	// Convert screen coordinates into modelspace coordinates
-	Vec3<double> &screenToModel(double x, double y, double z);
+	Vec3<double> &screenToModel(int x, int y, double z);
 	// Update transformation matrix
 	void setTransformationMatrix(Matrix &mat, Vec3<double> cellcentre);
